agrovision-azure-func/web/requirements.txt
azure-functions
azure-iot-hub
azure-cosmos